Added another macro. Navi select menu slides in post and pre entry. Added audio priority types. TODO: navi select menu finish. Add text box object for talking.

+++ BattleNetwork/bnField.cpp
@@ -122,7 +122,7 @@ void Field::Update(float _elapsed) {
   for (int y = 0; y < height; y++) {
     for (int x = 0; x < width; x++) {
       tiles[y][x]->Update(_elapsed);
-      entityCount += tiles[y][x]->entities.size();
+      entityCount += (int)tiles[y][x]->entities.size();
     }
   }
